:hammer: refactor: 인증/인가 설정 및 SSE 요청 URL 변경

See merge request s12-final/S12P31B204!99

+++ backend/src/main/java/com/s12p31b204/backend/controller/TestPaperController.java
@@ -67,16 +67,13 @@ public class TestPaperController {
              @AuthenticationPrincipal CustomOAuth2User user,
              HttpServletRequest request) {
         try {
-//            if(authorizationService.checkWorkBookAuthorization(user.getUserId(), workBookId)) {
-//                log.info("getting TestPapers...");
-//                List<TestPaperResponseDto> testPapers = testPaperService.findTestPaperByWorkBookId(workBookId);
-//                return ApiResponse.success(testPapers, "시험지 리스트 조회 성공", HttpStatus.OK, request.getRequestURI());
-//            } else {
-//                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
-//            }
-            log.info("getting TestPapers...");
-            List<TestPaperResponseDto> testPapers = testPaperService.findTestPaperByWorkBookId(workBookId);
-            return ApiResponse.success(testPapers, "시험지 리스트 조회 성공", HttpStatus.OK, request.getRequestURI());
+            if(authorizationService.checkWorkBookAuthorization(user.getUserId(), workBookId)) {
+                log.info("getting TestPapers...");
+                List<TestPaperResponseDto> testPapers = testPaperService.findTestPaperByWorkBookId(workBookId);
+                return ApiResponse.success(testPapers, "시험지 리스트 조회 성공", HttpStatus.OK, request.getRequestURI());
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (NoSuchElementException e) {
             return ApiResponse.failure(e.getMessage(), HttpStatus.BAD_REQUEST, request.getRequestURI());
         } catch (Exception e) {
@@ -92,16 +89,13 @@ public class TestPaperController {
             HttpServletRequest request
     ) {
         try {
-//            if(authorizationService.checkWorkBookAuthorization(user.getUserId(), createTestPaperRequestDto.getWorkBookId())) {
-//                log.info("creating TestPaper...");
-//                TestPaperResponseDto response = testPaperService.createTestPaper(createTestPaperRequestDto);
-//                return ApiResponse.success(response, "시험지 생성 성공", HttpStatus.CREATED, request.getRequestURI());
-//            } else {
-//                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
-//            }
-            log.info("creating TestPaper...");
-            TestPaperResponseDto response = testPaperService.createTestPaper(createTestPaperRequestDto, user.getUserId());
-            return ApiResponse.success(response, "시험지 생성 성공", HttpStatus.CREATED, request.getRequestURI());
+            if(authorizationService.checkWorkBookAuthorization(user.getUserId(), createTestPaperRequestDto.getWorkBookId())) {
+                log.info("creating TestPaper...");
+                TestPaperResponseDto response = testPaperService.createTestPaper(createTestPaperRequestDto, user.getUserId());
+                return ApiResponse.success(response, "시험지 생성 성공", HttpStatus.CREATED, request.getRequestURI());
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (NoSuchElementException e) {
             return ApiResponse.failure(e.getMessage(), HttpStatus.BAD_REQUEST, request.getRequestURI());
         } catch (Exception e) {
@@ -117,16 +111,13 @@ public class TestPaperController {
             HttpServletRequest request
     ) {
         try {
-//            if(authorizationService.checkWorkBookAuthorization(user.getUserId(), createTestPaperRequestDto.getWorkBookId())) {
-//                log.info("creating TestPaper...");
-//                TestPaperResponseDto response = testPaperService.createTestPaper(createTestPaperRequestDto);
-//                return ApiResponse.success(response, "시험지 생성 성공", HttpStatus.CREATED, request.getRequestURI());
-//            } else {
-//                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
-//            }
-            log.info("generating TestPaper...");
-            TestPaperResponseDto response = testPaperService.generateTestPaper(generateTestPaperRequestDto, user.getUserId());
-            return ApiResponse.success(response, "시험지 생성 성공", HttpStatus.CREATED, request.getRequestURI());
+            if(authorizationService.checkWorkBookAuthorization(user.getUserId(), generateTestPaperRequestDto.getWorkBookId())) {
+                log.info("generating TestPaper...");
+                TestPaperResponseDto response = testPaperService.generateTestPaper(generateTestPaperRequestDto, user.getUserId());
+                return ApiResponse.success(response, "시험지 생성 성공", HttpStatus.CREATED, request.getRequestURI());
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (NoSuchElementException e) {
             return ApiResponse.failure(e.getMessage(), HttpStatus.BAD_REQUEST, request.getRequestURI());
         } catch (Exception e) {
@@ -141,16 +132,13 @@ public class TestPaperController {
             @AuthenticationPrincipal CustomOAuth2User user,
             HttpServletRequest request) {
         try {
-//            if(authorizationService.checkTestPaperAuthorization(user.getUserId(), testPaperId)) {
-//                log.info("removing TestPaper...");
-//                testPaperService.removeTestPaper(testPaperId);
-//                return ApiResponse.success(null, "시험지 삭제 성공", HttpStatus.NO_CONTENT, request.getRequestURI());
-//            } else {
-//                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
-//            }
-            log.info("removing TestPaper...");
-            testPaperService.removeTestPaper(testPaperId);
-            return ApiResponse.success(null, "시험지 삭제 성공", HttpStatus.NO_CONTENT, request.getRequestURI());
+            if(authorizationService.checkTestPaperAuthorization(user.getUserId(), testPaperId)) {
+                log.info("removing TestPaper...");
+                testPaperService.removeTestPaper(testPaperId);
+                return ApiResponse.success(null, "시험지 삭제 성공", HttpStatus.NO_CONTENT, request.getRequestURI());
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (Exception e) {
             log.error(e.getMessage());
             return ApiResponse.failure("시험지 삭제 중 오류 발생", HttpStatus.INTERNAL_SERVER_ERROR, request.getRequestURI());
@@ -163,16 +151,14 @@ public class TestPaperController {
             @AuthenticationPrincipal CustomOAuth2User user,
             HttpServletRequest request) {
         try {
-//            if(authorizationService.checkTestPaperAuthorization(user.getUserId(), updateTestPaperRequestDto.getTestPaperId())) {
-//                log.info("modifying TestPaper...");
-//                TestPaperResponseDto response = testPaperService.updateTestPaper(updateTestPaperRequestDto);
-//                return ApiResponse.success(response, "시험지 수정 성공", HttpStatus.OK, request.getRequestURI());
-//            } else {
-//                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
-//            }
-            log.info("modifying TestPaper...");
-            TestPaperResponseDto response = testPaperService.updateTestPaper(updateTestPaperRequestDto);
-            return ApiResponse.success(response, "시험지 수정 성공", HttpStatus.OK, request.getRequestURI());
+            if(authorizationService.checkTestPaperAuthorization(user.getUserId(), updateTestPaperRequestDto.getTestPaperId())) {
+                log.info("modifying TestPaper...");
+                TestPaperResponseDto response = testPaperService.updateTestPaper(updateTestPaperRequestDto);
+                return ApiResponse.success(response, "시험지 수정 성공", HttpStatus.OK, request.getRequestURI());
+
+            } else {
+                return ApiResponse.failure("권한이 없습니다.", HttpStatus.FORBIDDEN, request.getRequestURI());
+            }
         } catch (NoSuchElementException e) {
             return ApiResponse.failure(e.getMessage(), HttpStatus.BAD_REQUEST, request.getRequestURI());
         } catch (Exception e) {
@@ -184,7 +170,6 @@ public class TestPaperController {
     @PostMapping("/convert-pdf")
     public void convertPdf(
             @RequestBody ConvertPdfRequestDto convertPdfRequestDto,
-            @AuthenticationPrincipal CustomOAuth2User user,
             HttpServletResponse response) {
         try {
             log.info("Convert TestPaper To PDF...");
